Move Cerberus utility functions to schema to avoid cyclic dependencies
