+++ Yank/schema/__init__.py
@@ -3,5 +3,5 @@ A set of tools for validating Cerberus-based schemas
 """
 
 from .validator import YANKCerberusValidator
-from .validator import generate_unknown_type_validator
-from .validator import type_to_cerberus_map
+from .validator import to_unit_validator
+from .validator import generate_unknown_type_validator, type_to_cerberus_map, generate_signature_schema
